requirements.txt
aiohttp==3.8.5
deepl==1.14.0
discord.py==2.2.3
gidgethub==5.2.1
httplib2==0.22.0
matplotlib==3.7.1
oauth2client==4.1.3
parsedatetime==2.6
pluralkit==1.1.5
pydrive==1.3.1
unidecode==1.3.6
